% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_ENMO}
\alias{add_ENMO}
\title{Add ENMO (Euclidean Norm Minus One) Metric}
\usage{
add_ENMO(day_matrix)
}
\arguments{
\item{day_matrix}{A data frame containing accelerometer data, including 'x', 'y', and 'z' columns.}
}
\value{
A data frame with the ENMO values added as a new column.
}
\description{
This function computes the ENMO metric, a measure of physical activity intensity,
and adds it as a new column to the dataset.
}
\examples{
day_matrix_with_enmo <- add_ENMO(day_matrix)
}
