% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{calibration}
\alias{calibration}
\title{Calibrate the GENEActiv Device}
\usage{
calibration(binfile)
}
\arguments{
\item{binfile}{The file path to the .bin file from the GENEActiv device.}
}
\value{
A numeric vector containing the calibration scale and calibration offset.
}
\description{
This function reads the .bin file from a GENEActiv device and returns
the calibration scale and offset values.
}
\examples{
calib <- calibration("path/to/binfile.bin")
}
