% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{compute_bouts}
\alias{compute_bouts}
\title{Compute Bouts of Activity}
\usage{
compute_bouts(D, freq, timestamps)
}
\arguments{
\item{D}{A data frame containing accelerometer data.}

\item{freq}{The sampling frequency.}

\item{timestamps}{A vector of timestamps corresponding to the data.}
}
\value{
A data frame with start and end times of each activity bout.
}
\description{
This function detects and computes periods of activity and inactivity based on accelerometer data.
}
\examples{
bouts <- compute_bouts(D, 100, timestamps)
}
