% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{compute_for_all}
\alias{compute_for_all}
\title{Compute Data for All Participants}
\usage{
compute_for_all(binfolder, outfolder)
}
\arguments{
\item{binfolder}{The path to the folder containing the .bin files.}

\item{outfolder}{The path to the output folder where results will be saved.}
}
\value{
A list containing data frames for all participants.
}
\description{
This function processes all .bin files in the specified folder, computes posture changes, and burstiness metrics.
}
\examples{
results <- compute_for_all("path/to/binfolder", "path/to/outfolder")
}
