% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract Metrics from a Raw Bout}
\usage{
extract_metrics(raw_bout, freq)
}
\arguments{
\item{raw_bout}{A data frame containing accelerometer data for a bout.}

\item{freq}{The sampling frequency.}
}
\value{
A list of metrics for the bout.
}
\description{
This function extracts various metrics from a raw bout of activity, such as volume and intensity.
}
