% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_interval}
\alias{get_interval}
\title{Get Data for a Specified Time Interval}
\usage{
get_interval(v_bin_data, binfile, start, end, calib, freq)
}
\arguments{
\item{v_bin_data}{The binary data read from the GENEActiv device.}

\item{binfile}{The file path to the .bin file from the GENEActiv device.}

\item{start}{The start time for the data extraction.}

\item{end}{The end time for the data extraction.}

\item{calib}{A calibration vector obtained from the `calibration()` function.}

\item{freq}{The frequency of measurements in Hz.}
}
\value{
A list containing the following elements:
\item{DataFrame}{The extracted accelerometer data.}
\item{Date}{The date of the midpoint of the extracted data.}
\item{WeekDay}{The weekday of the extracted date.}
\item{DayNo}{The day number within the study.}
\item{NonWear}{The number of non-wear hours.}
}
\description{
This function reads accelerometer data from the GENEActiv device for a specified time interval,
applies the calibration, and calculates the number of non-wear hours.
}
\examples{
interval_data <- get_interval(v_bin_data, "path/to/binfile.bin", "2023-01-01", "2023-01-02", calib, 100)
}
