% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Get Raw Data for Bouts}
\usage{
get_raw_data(D, bouts, freq)
}
\arguments{
\item{D}{A data frame containing accelerometer data.}

\item{bouts}{A data frame with start and end times for each bout.}

\item{freq}{The sampling frequency.}
}
\value{
A list of data frames for each bout.
}
\description{
This function extracts raw accelerometer data for specified activity bouts.
}
