% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{measure_burstiness}
\alias{measure_burstiness}
\title{Measure Burstiness for All Participants}
\usage{
measure_burstiness(results)
}
\arguments{
\item{results}{The list containing data frames of all participants' data.}
}
\value{
A data frame containing burstiness metrics for all participants.
}
\description{
This function calculates burstiness metrics for all participants.
}
\examples{
burstiness_results <- measure_burstiness(results)
}
