% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{non_wear_window}
\alias{non_wear_window}
\title{Detect Non-Wear Windows}
\usage{
non_wear_window(window)
}
\arguments{
\item{window}{A data frame containing accelerometer data and temperature values.}
}
\value{
An integer status indicating non-wear (0) or wear (1).
}
\description{
This function identifies periods where the GENEActiv device was likely not worn based on
temperature and movement data.
}
\examples{
status <- non_wear_window(window_data)
}
