% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{run_full_analysis}
\alias{run_full_analysis}
\title{Run Full Analysis on GENEActiv Data}
\usage{
run_full_analysis(binfolder, outfolder)
}
\arguments{
\item{binfolder}{The path to the folder containing the .bin files.}

\item{outfolder}{The path to the output folder where results will be saved.}
}
\value{
A data frame containing burstiness metrics for all subjects.
}
\description{
This function processes all GENEActiv .bin files in the specified folder, computes burstiness metrics for each person,
and saves the results in the output folder.
}
\examples{
burstiness_metrics <- run_full_analysis("D:/Downloads/p004", "D:/Downloads/out/conametest")
}
