% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{SedUp}
\alias{SedUp}
\title{Posture Detection Using SedUp}
\usage{
SedUp(y, fs, option)
}
\arguments{
\item{y}{A numeric vector of accelerometer data (single axis).}

\item{fs}{The sampling frequency of the data.}

\item{option}{Integer specifying which posture detection option to use (1 or 2).}
}
\value{
A numeric vector where 1 indicates standing and 0 indicates sedentary behavior.
}
\description{
This function detects periods of sedentary behavior or standing using a logistic regression model.
}
\examples{
standing_periods <- SedUp(y_data, 100, 1)
}
