% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{wind_SD}
\alias{wind_SD}
\title{Compute Windsorized Standard Deviation}
\usage{
wind_SD(vect)
}
\arguments{
\item{vect}{A numeric vector of data.}
}
\value{
The Windsorized standard deviation of the input vector.
}
\description{
This function calculates the Windsorized standard deviation, reducing the influence of outliers.
}
\examples{
wind_sd <- wind_SD(data_vector)
}
